Time: 2 ms (71.63%), Space: 6.3 MB (5.31%) - LeetHub

+++ 0441-arranging-coins/0441-arranging-coins.cpp
@@ -4,7 +4,7 @@ public:
         int l = 1, r = 65535, mid;
         while (l <= r) {
             mid = (l + r) / 2;
-            int tmp = ((long long)mid * (mid + 1)) / 2;
+            long long tmp = ((long long)mid * (mid + 1)) / 2;
             if (tmp > n) {
                 r = mid - 1;
             }
